requirements.txt
streamlit
python-dotenv
fastapi
uvicorn[standard]
requests
pydantic
